Delete has now confirmation with translation

+++ app/controllers/books/index.js
@@ -1,9 +1,12 @@
 import Controller from '@ember/controller';
+import { inject } from '@ember/service';
 
 export default Controller.extend({
+  i18n: inject(),
+
   actions: {
     deleteBook(book) {
-      book.destroyRecord();
+      if(confirm(this.get('i18n').t('delete.confirm'))) book.destroyRecord();
     }
   }
 });
